requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
protobuf==3.20.1
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
werkzeug<2.1
python-decouple==3.4
flask_mysqldb
flask_api
ecdsa==0.17.0 
pyasn1==0.4.8 
python-jose==3.3.0 
rsa==4.8
cryptography==37.0.1
bcrypt==3.2.2
pytest==7.1.2
apispec==5.2.1
apispec-webframeworks==0.5.2
marshmallow==3.15.0
Flask-Testing==0.8.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.1